Integrated CSmartMean class code into sl_tools.h and sl_tools.cpp
